+++ CHANGELOG.md
@@ -3,9 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+# [3.47.0] - 31-08-2021
+### Added
+- Css assets of client async fragments will be downloaded asynchronously. Also, the styles of the client async fragment with the critical-css property will be in the source code.
 # [3.46.0] - 23-11-2020
-### Addd
+### Added
 - main and partial placeholders seperated for client-async mode, now placeholder payload can be string or an object
+# [3.45.0] - 23-11-2020
+### Added
+- Implemented on demand client async feature.Fragments can be loaded externally according to user action.
 # [3.33.0] - 23-11-2020
 ### Fix
 - Fix for redirectionEnabled attribute to allow ssr redirection.